requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=1.1.0=cpu
asn1crypto=0.24.0=py37_0
blas=1.0=mkl
bzip2=1.0.6=hfa6e2cd_5
ca-certificates=2019.5.15=0
certifi=2019.6.16=py37_0
cffi=1.12.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1
conda=4.6.14=py37_0
conda-package-handling=1.3.10=py37_0
cryptography=2.7=py37h7a1dbc1_0
cudatoolkit=9.0=1
freetype=2.9.1=ha9979f8_1
future=0.17.1=py37_1000
icc_rt=2019.0.0=h0cc432a_1
idna=2.8=py37_0
intel-openmp=2019.4=245
jpeg=9b=hb83a4c4_2
libarchive=3.3.3=h0643e63_5
libiconv=1.15=h1df5818_7
libpng=1.6.37=h2a8f88b_0
libtiff=4.0.10=h2929a5b_1001
libxml2=2.9.9=h464c3ec_0
lz4-c=1.8.1.2=h2fa13f4_0
lzo=2.10=h6df0209_2
menuinst=1.4.16=py37he774522_0
mkl=2019.4=245
mkl_fft=1.0.12=py37h14836fe_0
mkl_random=1.0.2=py37h343c172_0
ninja=1.9.0=py37h74a9793_0
numpy=1.16.4=py37h19fb1c0_0
numpy-base=1.16.4=py37hc3f5095_0
olefile=0.46=py37_0
openssl=1.1.1=he774522_0
pillow=6.0.0=py37hdc69c19_0
pip=19.1.1=py37_0
pycosat=0.6.3=py37hfa6e2cd_0
pycparser=2.19=py37_0
pyopenssl=19.0.0=py37_0
pysocks=1.7.0=py37_0
pytelegrambotapi=3.6.6=pypi_0
python=3.7.3=h8c8aaf0_1
python-libarchive-c=2.8=py37_6
python-telegram-bot=11.1.0=py_0
pytorch=1.1.0=py3.7_cuda90_cudnn7_1
pywin32=223=py37hfa6e2cd_1
requests=2.22.0=py37_0
ruamel_yaml=0.15.46=py37hfa6e2cd_0
scipy=1.2.1=py37h29ff71c_0
setuptools=41.0.1=py37_0
six=1.12.0=py37_0
sqlite=3.28.0=he774522_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.3.0=py37_cu90_1
tqdm=4.32.1=py_0
ujson=1.35=py37hfa6e2cd_1001
urllib3=1.24.2=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.15.26706=h3a45250_4
wheel=0.33.4=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
